int64 -> uint64 in some places. Factor out PrepareForRefund and testutil.ChunkStore.

+++ cmd/tredd/filechunks.go
@@ -7,11 +7,11 @@ import (
 
 type fileChunkStore struct {
 	filename  string
-	chunksize int64
-	size      int64 // size of file in bytes
+	chunksize uint64
+	size      uint64 // size of file in bytes
 }
 
-func newFileChunkStore(filename string, chunksize int64) (*fileChunkStore, error) {
+func newFileChunkStore(filename string, chunksize uint64) (*fileChunkStore, error) {
 	result := &fileChunkStore{
 		filename:  filename,
 		chunksize: chunksize,
@@ -22,7 +22,7 @@ func newFileChunkStore(filename string, chunksize int64) (*fileChunkStore, error
 	} else if err != nil {
 		return nil, err
 	} else {
-		result.size = info.Size()
+		result.size = uint64(info.Size())
 	}
 	return result, nil
 }
@@ -37,18 +37,18 @@ func (s *fileChunkStore) Add(bits []byte) error {
 	if err != nil {
 		return err
 	}
-	s.size += int64(len(bits))
+	s.size += uint64(len(bits))
 	return nil
 }
 
-func (s *fileChunkStore) Get(index int64) ([]byte, error) {
+func (s *fileChunkStore) Get(index uint64) ([]byte, error) {
 	f, err := os.Open(s.filename)
 	if err != nil {
 		return nil, err
 	}
 	defer f.Close()
 
-	_, err = f.Seek(index*s.chunksize, os.SEEK_SET) // TODO: range check
+	_, err = f.Seek(int64(index*s.chunksize), os.SEEK_SET) // TODO: range check
 	if err != nil {
 		return nil, err
 	}
@@ -58,14 +58,14 @@ func (s *fileChunkStore) Get(index int64) ([]byte, error) {
 	n, err := io.ReadFull(f, result)
 	if err == io.ErrUnexpectedEOF {
 		// Partial chunk allowed only at EOF.
-		if index*s.chunksize+int64(n) == s.size {
+		if index*s.chunksize+uint64(n) == s.size {
 			return result[:n], nil
 		}
 	}
 	return result[:n], err
 }
 
-func (s *fileChunkStore) Len() (int64, error) {
+func (s *fileChunkStore) Len() (uint64, error) {
 	n, r := s.size/s.chunksize, s.size%s.chunksize
 	if r > 0 {
 		n++
