+++ serve.go
@@ -18,7 +18,7 @@ func Serve(w io.Writer, r io.Reader, key [32]byte) ([]byte, error) {
 		hasher   = sha256.New()
 	)
 
-	for index := int64(0); ; index++ {
+	for index := uint64(0); ; index++ {
 		var chunk [ChunkSize]byte
 		n, err := io.ReadFull(r, chunk[:])
 		if err == io.EOF {
@@ -27,7 +27,7 @@ func Serve(w io.Writer, r io.Reader, key [32]byte) ([]byte, error) {
 			return nil, errors.Wrapf(err, "reading clear chunk %d", index)
 		}
 
-		prefixedClearChunk := PrefixChunk(uint64(index), chunk[:n])
+		prefixedClearChunk := PrefixChunk(index, chunk[:n])
 		var clearChunkHash [32]byte
 		merkle.LeafHash(hasher, clearChunkHash[:0], prefixedClearChunk)
 		_, err = w.Write(clearChunkHash[:])
@@ -41,7 +41,7 @@ func Serve(w io.Writer, r io.Reader, key [32]byte) ([]byte, error) {
 			return nil, errors.Wrapf(err, "writing cipher chunk %d", index)
 		}
 
-		prefixedCipherChunk := PrefixChunk(uint64(index), chunk[:n])
+		prefixedCipherChunk := PrefixChunk(index, chunk[:n])
 		cipherMT.Add(prefixedCipherChunk)
 	}
